fix: accept decimal currency amounts in check-in settings
- 签到奖励设置 min/max 曾用 z.coerce.number().int()，但字段为显示货币金额（如 0.5），小数必被拒绝并报 Invalid input
- 改为 .min(0) 允许小数；整数 quota 换算仍由 parseQuotaFromDollars 完成（0.5 -> 250000, 0.8 -> 400000 @ quotaPerUnit=500000）
- 新增回归测试：checkin-settings-section.test.tsx（提交 0.5/0.8 成功且无 Invalid input）
- 验证：vitest 1 passed；bun run typecheck (tsgo -b) exit 0

本变更由 AI 辅助生成并人工复核。

+++ web/src/features/system-settings/general/checkin-settings-section.tsx
@@ -59,8 +59,10 @@ import { useUpdateOption } from '../hooks/use-update-option'
 
 const schema = z.object({
   enabled: z.boolean(),
-  minQuota: z.coerce.number().int().min(0),
-  maxQuota: z.coerce.number().int().min(0),
+  // 表单字段是显示货币金额（如 ¥0.5 = 0.5），必须允许小数；
+  // 内部 quota 的整数换算在 parseQuotaFromDollars 内完成。
+  minQuota: z.coerce.number().min(0),
+  maxQuota: z.coerce.number().min(0),
 })
 
 type Values = z.infer<typeof schema>
